+++ src/components/Livros/livros.scss
@@ -12,6 +12,7 @@
   width: 100%;
   height: 200px;
   object-position: 0 50%;
+  filter: blur(20px);
 }
 
 .foto-do-livro img {
